made updates to code, including splitting "EEG_feature_extractor.mlx" into "EEG_Analysis.mlx" and "EEG_PreProcess.mlx". From this point forward, "EEG_feature_extractor.mlx" will not be updated. Refer to the other two .mlx files for up to date code.
created function "spectral_bandpower.mlx" to output the spectral power value for specified frequency bands. 

TODO: modify "spectral_bandpower.mlx" to accommodate frequency band ranges for which values in the array do not exist. for example, if you input [30 100] but the data only exists only to 60Hz, change frequency band to [30 60].

TODO: finish modifying portion of "EEG_Analysis.mlx" to replace modes within a window with NaN. This is to remove noise from the dataset