reorganized "EEG_Analysis.mlx" a bit for readability
added "windowed_data.mlx", a function to break apart data into a 3D array of windows, in our case, 15 second windows. this is used when putting data into GMM

TODO: modify "spectral_bandpower.mlx" to accommodate frequency band ranges for which values in the array do not exist. for example, if you input [30 100] but the data only exists only to 60Hz, change frequency band to [30 60].

TODO: finish modifying portion of "EEG_Analysis.mlx" to replace modes within a window with NaN. This is to remove noise from the dataset